+++ pkg/apis/etcd/v1beta3/doc.go
@@ -16,4 +16,4 @@ limitations under the License.
 
 // +k8s:deepcopy-gen=package
 // +groupName=etcd.database.coreos.com
-package v1beta2
+package v1beta3