+++ pkg/util/k8sutil/k8sutils_test.go
@@ -17,7 +17,7 @@ package k8sutil
 import (
 	"testing"
 
-	api "github.com/coreos/etcd-operator/pkg/apis/etcd/v1beta2"
+	api "github.com/coreos/etcd-operator/pkg/apis/etcd/v1beta3"
 )
 
 func TestDefaultBusyboxImageName(t *testing.T) {