+++ test/e2e/pv_test.go
@@ -18,8 +18,8 @@ import (
 	"os"
 	"testing"
 
-	api "github.com/coreos/etcd-operator/pkg/apis/etcd/v1beta2"
-	"k8s.io/api/core/v1"
+	api "github.com/coreos/etcd-operator/pkg/apis/etcd/v1beta3"
+	v1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 
 	"github.com/coreos/etcd-operator/test/e2e/e2eutil"